@@ -18,26 +18,24 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: '>=3.3.0 <4.0.0'
+  flutter: '>=3.19.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  made_with_serverpod: ^1.0.0
-  pixels: ^1.0.1
+  serverpod_flutter: 2.1.3
   pixorama_client:
     path: ../pixorama_client
-  serverpod_flutter:
-    path: ../vendor/serverpod/packages/serverpod_flutter
-  url_strategy: ^0.2.0
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.5
+  pixels: ^1.0.1
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: '>=3.0.0 <5.0.0'
   flutter_test:
     sdk: flutter
 
