@@ -1,37 +1,37 @@
 /* AUTOMATICALLY GENERATED CODE DO NOT MODIFY */
 /*   To generate run: "serverpod generate"    */
 
-// ignore_for_file: library_private_types_in_public_api
-// ignore_for_file: public_member_api_docs
 // ignore_for_file: implementation_imports
+// ignore_for_file: library_private_types_in_public_api
+// ignore_for_file: non_constant_identifier_names
+// ignore_for_file: public_member_api_docs
+// ignore_for_file: type_literal_in_constant_pattern
+// ignore_for_file: use_super_parameters
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:serverpod_client/serverpod_client.dart' as _i1;
 import 'dart:typed_data' as _i2;
 
-class ImageData extends _i1.SerializableEntity {
-  ImageData({
-    this.id,
+abstract class ImageData implements _i1.SerializableModel {
+  ImageData._({
     required this.pixels,
     required this.width,
     required this.height,
   });
 
-  factory ImageData.fromJson(
-    Map<String, dynamic> jsonSerialization,
-    _i1.SerializationManager serializationManager,
-  ) {
+  factory ImageData({
+    required _i2.ByteData pixels,
+    required int width,
+    required int height,
+  }) = _ImageDataImpl;
+
+  factory ImageData.fromJson(Map<String, dynamic> jsonSerialization) {
     return ImageData(
-      id: serializationManager.deserialize<int?>(jsonSerialization['id']),
-      pixels: serializationManager
-          .deserialize<_i2.ByteData>(jsonSerialization['pixels']),
-      width: serializationManager.deserialize<int>(jsonSerialization['width']),
-      height:
-          serializationManager.deserialize<int>(jsonSerialization['height']),
+      pixels: _i1.ByteDataJsonExtension.fromJson(jsonSerialization['pixels']),
+      width: jsonSerialization['width'] as int,
+      height: jsonSerialization['height'] as int,
     );
   }
-
-  int? id;
 
   _i2.ByteData pixels;
 
@@ -39,13 +39,47 @@ class ImageData extends _i1.SerializableEntity {
 
   int height;
 
+  ImageData copyWith({
+    _i2.ByteData? pixels,
+    int? width,
+    int? height,
+  });
   @override
   Map<String, dynamic> toJson() {
     return {
-      'id': id,
-      'pixels': pixels,
+      'pixels': pixels.toJson(),
       'width': width,
       'height': height,
     };
+  }
+
+  @override
+  String toString() {
+    return _i1.SerializationManager.encode(this);
+  }
+}
+
+class _ImageDataImpl extends ImageData {
+  _ImageDataImpl({
+    required _i2.ByteData pixels,
+    required int width,
+    required int height,
+  }) : super._(
+          pixels: pixels,
+          width: width,
+          height: height,
+        );
+
+  @override
+  ImageData copyWith({
+    _i2.ByteData? pixels,
+    int? width,
+    int? height,
+  }) {
+    return ImageData(
+      pixels: pixels ?? this.pixels.clone(),
+      width: width ?? this.width,
+      height: height ?? this.height,
+    );
   }
 }
