@@ -1,28 +1,31 @@
 /* AUTOMATICALLY GENERATED CODE DO NOT MODIFY */
 /*   To generate run: "serverpod generate"    */
 
-// ignore_for_file: library_private_types_in_public_api
-// ignore_for_file: public_member_api_docs
 // ignore_for_file: implementation_imports
+// ignore_for_file: library_private_types_in_public_api
+// ignore_for_file: non_constant_identifier_names
+// ignore_for_file: public_member_api_docs
+// ignore_for_file: type_literal_in_constant_pattern
+// ignore_for_file: use_super_parameters
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:serverpod_client/serverpod_client.dart' as _i1;
 
-class ImageUpdate extends _i1.SerializableEntity {
-  ImageUpdate({
+abstract class ImageUpdate implements _i1.SerializableModel {
+  ImageUpdate._({
     required this.pixelIndex,
     required this.colorIndex,
   });
 
-  factory ImageUpdate.fromJson(
-    Map<String, dynamic> jsonSerialization,
-    _i1.SerializationManager serializationManager,
-  ) {
+  factory ImageUpdate({
+    required int pixelIndex,
+    required int colorIndex,
+  }) = _ImageUpdateImpl;
+
+  factory ImageUpdate.fromJson(Map<String, dynamic> jsonSerialization) {
     return ImageUpdate(
-      pixelIndex: serializationManager
-          .deserialize<int>(jsonSerialization['pixelIndex']),
-      colorIndex: serializationManager
-          .deserialize<int>(jsonSerialization['colorIndex']),
+      pixelIndex: jsonSerialization['pixelIndex'] as int,
+      colorIndex: jsonSerialization['colorIndex'] as int,
     );
   }
 
@@ -30,11 +33,41 @@ class ImageUpdate extends _i1.SerializableEntity {
 
   int colorIndex;
 
+  ImageUpdate copyWith({
+    int? pixelIndex,
+    int? colorIndex,
+  });
   @override
   Map<String, dynamic> toJson() {
     return {
       'pixelIndex': pixelIndex,
       'colorIndex': colorIndex,
     };
+  }
+
+  @override
+  String toString() {
+    return _i1.SerializationManager.encode(this);
+  }
+}
+
+class _ImageUpdateImpl extends ImageUpdate {
+  _ImageUpdateImpl({
+    required int pixelIndex,
+    required int colorIndex,
+  }) : super._(
+          pixelIndex: pixelIndex,
+          colorIndex: colorIndex,
+        );
+
+  @override
+  ImageUpdate copyWith({
+    int? pixelIndex,
+    int? colorIndex,
+  }) {
+    return ImageUpdate(
+      pixelIndex: pixelIndex ?? this.pixelIndex,
+      colorIndex: colorIndex ?? this.colorIndex,
+    );
   }
 }
